+++ src/PhpGenerator/Printer.php
@@ -48,7 +48,7 @@ class Printer
 		foreach ($closure->getUses() as $param) {
 			$uses[] = ($param->isReference() ? '&' : '') . '$' . $param->getName();
 		}
-		$useStr = strlen($tmp = implode(', ', $uses)) > Helpers::WRAP_LENGTH && count($uses) > 1
+		$useStr = strlen($tmp = implode(', ', $uses)) > Dumper::WRAP_LENGTH && count($uses) > 1
 			? "\n" . $this->indentation . implode(",\n" . $this->indentation, $uses) . "\n"
 			: $tmp;
 
@@ -216,7 +216,7 @@ class Printer
 
 	protected function dump($var): string
 	{
-		return Helpers::dump($var);
+		return Dumper::dump($var);
 	}
 
 
@@ -254,7 +254,7 @@ class Printer
 				. ($param->hasDefaultValue() && !$variadic ? ' = ' . $this->dump($param->getDefaultValue()) : '');
 		}
 
-		return strlen($tmp = implode(', ', $params)) > Helpers::WRAP_LENGTH && count($params) > 1
+		return strlen($tmp = implode(', ', $params)) > Dumper::WRAP_LENGTH && count($params) > 1
 			? "(\n" . $this->indentation . implode(",\n" . $this->indentation, $params) . "\n)"
 			: "($tmp)";
 	}
